Update patch set 2
Patch Set 2:

Summary:

This and dependent CLs, userdebug, verity off, 20 complaints about missing verity device devices reduced to 6. Blocked from reducing it to 3 b/c we first check if overlayfs is practical to instantiate scratch partition, then another 3 b/c we actually try to use the scratch partition.

Can reduce it to 2 if we merge the partition list (as we formerly proposed), this time we can make an effort to preserve the errno(?). Worth it because of how precious the kernel logs are _before_ logd.klogd starts up.

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>