Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I still don't know why we didn't just refactor the original function to allow taking a single fstab_rec.  We know we need to handle / as a special case for system as root, but everything else would go smoothly through such a function.

Because the rework to support fstab_rec is useless here because there is no fstab_rec, and reworking was quite cumbersome to pass fstab_rec in all cases because of the additional 4+ parameters also required (but could be wrapped).

Here we also lack the ability to determine if avb is supported or not, but we infer it from checking both "system" and "vroot".

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0