Update patch set 3
Patch Set 3:

first_stage_mount.cpp doesn't need this though, right?  The only place that we need to worry about this, we action based on the presence of a "/system" mount point.

It's just the overlayfs mount in the first stage that does?  It can likely safely assume that if there is no "/system" mount point that it is system-as-root, as long as it's being called from first stage mount.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>