Update patch set 1
Patch Set 1:

Actually my question is "can we make it an error in this project". I suppose you're saying that if we are able to fix these warnings tree-wide, there's no point in doing that because it will become an error everywhere?

LGTM either way.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>