Update patch set 2
Patch Set 2: Presubmit-Ready+1

(2 comments)

After our f2f chat I implemented the changes we discussed. Now APEX will fail soon in the case of no-streaming, fastdeploy or install-multiple with a clear error message. For the streaming case, an API check will ensure immediate failure. Since the APEX feature will really be available in API 29, not released yet, for now we check if the API level is 28, and we let "cmd" fail if the --apex flag is not supported on the device.

> Patch Set 1:
> 
> (1 comment)
> 
> Thought about this in a bit more detail. For this change you need to test at least 5 scenarios:
> 
> install with --streaming (equivalent to your current tests)

Tested:
- if API >= 28 and apex are supported in PackageManager, succeed.
- if API >= 28 and apex not supported by PackageManager, fails with error from PackageManager saying that --apex flag is not supported
- if API < 28, fails immediately

> install with --no-streaming (forces the default behaviour for <L devices)

- if an apex package is passed as argument, fails with message: "adb: APEX packages are only compatible with Streamed Install, exiting."

> install with -r --fastdeploy --no-streaming (general test for fast deploy)

- if an apex package is passed as argument, fails with message: "adb: APEX packages are only compatible with Streamed Install, exiting."

> install with -r --fastdeploy --streaming (<L test for fast deploy)

Similar to --streaming test, except it outputs a warning that fastdeploy has been disabled, e.g.:

"Performing Streamed Install
APEX is not compatible with Fast Deploy, ignoring --fastdeploy setting.
Success"

> install-multiple (does not support any of the above options so a single test is sufficient)

No need to support install-multiple for APEX, so in case we find an apex file we fail immediately.

Patch-set: 2
Reviewer: Gerrit User 1305591 <1305591@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1