Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> good idea or bad idea? (still following on from the trouble NO_ERROR caused for idries the other week.)
> 
> i'd have to move aapt from NO_ERROR to OK, and maybe some other bits of code that are built for Windows, but that's doable. _if_ we think this is worth it. given that i don't think we can remove NO_ERROR completely for all platforms (~5k files), i'm not sure. opinions?

It might be possible to do a broader cleanup of NO_ERROR in the future with other OWNERS of the code (maybe as a fixit of some sort? maybe even alongside some other cleanup work?). Does this patch cause any problems with -Wenum-conversion or -Wenum-compare? If not, then I think that the results definitely look better.

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>