Update patch set 1
Patch Set 1: Verified+1

> Why though?  I was iffy about allowing f2fs in the first place.  Do we really have a use case for it?  Do we have a use case for this alternative mount scenario?  It seems like a code path that's going to never be used and just bitrot.

f2fs is system group preference as it reduces wear on the flash devices.

It just so happens that in fastboot (or recovery mode) we can not use the existence of the formatting executable to determine the default filesystem. Even if the default is f2fs for the device, we can not guarantee that the filesystem is generated in the context of recovery mode, so the tool _is_ _not_ there on our pixel devices. Thus we have a preference for f2fs _if_ we can make the filesystem on the normal boot, all we need is the capability to mount the formatted filesystem in fastbootd regardless of the type.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Presubmit-Ready=0