Update patch set 1
Patch Set 1:

The tests that seem to have failed pass locally on both 32 and 64-bit (and they seem impossible to fail without everything being broken).

This seems like it might be a bug in the test runner? There's a bunch of exceptions in the shard 2 host log where it's trying to peek into an empty stack.

Patch-set: 1