Update patch set 1
Patch Set 1:

This change should not have an effect, there is no stack checking going on as part of that blacklist. If it does have an effect, then we have to view this as a _bug_ in llkd.

stack default is leaking? Will investigate ...

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>