Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Why isn't this just a coding problem on the other end?
> > 
> > Eval is called a number of times, so this check seems like it's not very helpful and would change performance.
> 
> Without this check, a malicious binary could crash us by having a 64bit Dwarf entry but pretending to be 32 bit regarding the registers it sends.
> 
> Is there some better place to put this up the callchain?

I think there is a better place for this, but let me think about it for a bit.

I think it could be done at after an elf creation, that you verify that the arch matches the register object arch. I need to see if there is one single place that can occur.

Patch-set: 2
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>