adbd: Fix adb remount when using dynamic partitions.
This patch addresses two issues when using dynamic partitions. First,
the blk_device field in fstab_rec will be incorrect if referring to a
device-mapper node (fs_mgr_update_logical_partition must be called).
Second, the device will appear to be system-as-root, but the fstab entry
is "/system" rather than "/".

To simplify this a bit, remount_partition is now explicitly called with
"/system" when not system-as-root.

Bug: N/A
Test: adb remount on device with dynamic partitions
Change-Id: I48b379db49d2def555ea49888c4994fbcf501672