Update patch set 1
Patch Set 1:

(2 comments)

I gotta think through this a bit more.  I didn't quite realize when I pushed this that for system-as-root, the default fstab is always present, so it'll fail to mount anything (since the flag isn't there).  We do need a backwards compatible solution, so something else will need to be done.

We could have a separate 'first stage fstab' but I want to avoid that if possible.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0