fastboot: Query the name of the super partition.
This patch adds a new variable, "super-partition-name", to query the
name of the super partition (with a slot suffix if it has one). The
fastboot flashing tool has been updated to query this variable.

Additionally, the set_active call has been moved from the end of the
flashing sequence to the beginning. This ensures that we pick the
correct super partition when flashing.

Finally, since the super partition name can no longer be determined
without fastbootd, the presence of super_empty.img is used to test for
dynamic partition support rather than the presence of a super partition.

Bug: 116802789
Test: fastboot flashall
Change-Id: If830768eba6de7f31ac3183c64167fae973c77a4