Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> if i add ziparchive-tests, i get this error:
> 
> WARNING: Please add ziparchive-tests to either suite: ['device-tests',
> 'general-tests'] for this TEST_MAPPING file to work with TreeHugger.
> 
> but i don't get that if i add, say, init_tests or debuggerd_test.
> what's this about?

https://cs.corp.google.com/android/system/core/debuggerd/Android.bp?q=debuggerd_test&dr=C&l=237

This is about packaging on the builders to move the artifacts to run on the test servers. You get 2 options: device-tests, meaning the test is compiled against an exact CPU, according to ccross/dwillemsen this is likely preferred for gtests. The other option is general-tests, meaning it is just compiled against a generic abi, preferred for java tests.

Patch-set: 2
Reviewer: Gerrit User 1083173 <1083173@85c56323-6fa9-3386-8a01-6480fb634889>