Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > (1 comment)
> 
> Dear Josh,
> 
> pipe_write.release():
> will set pipe_write.get()=-1; then ~unique_fd_impl()->reset()-> will not close this write fd for value_ == -1
> so we need to add close fd when unpack_thread_fd is not -1.
> 
> also, i have make a test, 
> and found it will be always fd leak when modify code as below：
>  iterate_sliblings(), dismiss if ,  let’s  always return error.

fd isn't pipe_write.release(), though. It's the value that some other thread has put into trace_output. That code is run when the compare_exchange fails because there is already an fd in trace_output. We need to close the file descriptor we tried to put into trace_output, not the one that is already there, because someone will close it eventually.

I think the proper fix is to do this instead:

        uint64_t expected = pack_thread_fd(-1, -1);
        int sent_fd = pipe_write.release();
        if (!trace_output.compare_exchange_strong(expected,
                                                  pack_thread_fd(tid, sent_fd))) {
          auto [tid, fd] = unpack_thread_fd(expected);
          async_safe_format_log(ANDROID_LOG_ERROR, "libc",
                                "thread %d is already outputting to fd %d?", tid, fd);
          close(sent_fd);
          return false;
        }

Patch-set: 2
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>