Update patch set 2
Patch Set 2: Presubmit-Ready+1

Just going through some old emails and catching up to this.

This change is strictly better than before, right?  Before we would drop the contents and not free the old pointer vs after we drop the contents but also free the old pointer?  Might as well take it in that case and file a tracking bug for the future?

Alternatively, how likely are we to run out of memory in this condition?  Is it likely enough that having recovery code going to be a worthwhile adventure?  None of the logging code should be coming anywhere close to running out of memory, right?  Maybe we just take this and move on?

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1