Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> why isn't that shared library dependency working? we already declare
> it in the cc_test:
> 
>     required: [
>         "libdemangle",
>     ],

+Dan/Colin, who are working on shared libraries for host-tests.

Atest and Test Mapping for device gtests so far has been assuming that they've been statically linked (which would be the easy fix here) but seems we have a gap here in supporting dynamically linked gtests.

Colin, thoughts here? I think we need to package the *.so into the testcase directory and push it over similarly to test data.

Dan will your current packaging work for host gtests help here or is this out of scope?

Patch-set: 1