fastbootd: Rewrite partition names when using update-super on retrofit A/B devices.
On retrofit A/B devices, the super partition will only span across a
single slot, not both slots. This means the LpMetadataBlockDevice list
must be slot-suffixed. Unfortunately this is difficult to do without
building two separate super images, or adding a flag to
LpMetadataGeometry and requiring all consumers of LpMetadata to append
slot suffixes.

The former is a lot of build system plumbing for little gain. The latter
is slightly more appealing, but only if consumers acquired MetadataBuilders
through fs_mgr rather than liblp (since liblp can't call
fs_mgr_get_slot_suffix).

As a simple fix, we can instead modify the block device list within
fastbootd. This could be done on either the host or device side, but is
done on the device side here since host fastboot cannot yet link to
liblp on Windows.

Note that this solves the "fastboot flashall" flow, but not the
"fastboot flash super" flow. However, retrofit devices do not have a
physical super partition, so this use case will need additional fixes
anyway.

Bug: 116802789
Test: manual test
Change-Id: Id5e720c1298e085e93c7478f4316bcfafd49433b