watchdog: slow down for serial console
If the serial console is active, increase the interval by an order
of magnitude, as the serial console can slow the system down, and
will consume real time CPU cycles in the CPU when flushing content.
With console enabled, also assume a developer will appreciate the
longer watchdog to enable a quick investigation.

Test: manual, check dmesg for watchdogd messages
Bug: 118712403
Change-Id: Idd93403b4e4d67f5dee1670f1ffaa3f334e92a6c