Update patch set 1
Patch Set 1: Code-Review+2

although part of me wonders whether it makes sense to have a "don't do that" overload of the "real" name and then a "does the right thing" function with almost the exact same name...

but at least -- going back to my "we should have more actionable error messages for the overloads" from yesterday -- this is something we can say in those errors. something like "fdopen function takes ownership of the fd; if that's not acceptable, use dup(2); if that is acceptable, use Fdopen which also handles the failure case".

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2