Update patch set 1
Patch Set 1:

Wait, does this mean that if I flash the 'a' slot, then I can't boot into what used to be on the 'b' slot anymore?  That seems wrong; we should probably think about what to do for update-super too, given I missed that point in the review.

Basically, would this be supported:
1) Device booting normally with slot a
2) Switch to slot b, flash image, fails to boot
3) Switch back to slot a, and it boots normally again

Assuming the original image and the image flashed in 2) have a different layout.

Alternatively thought as, like we've been talking about supporting 'a' vs 'b' vs 'all' for slot override, we should only update the metadata for the given slot(s) that fastboot is set to update.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>