Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Wait, does this mean that if I flash the 'a' slot, then I can't boot into what used to be on the 'b' slot anymore?  That seems wrong; we should probably think about what to do for update-super too, given I missed that point in the review.

Indeed I broke that in the update-super patch (more on that below). For this patch, it simply improves the existing scenario. The following sequence is currently broken:

  fastboot set_active a
  fastboot flash system_a ...
  fastboot set_active b
  fastboot flash system_a ...

We currently wind up with two different system_a partitions, occupying different extents, which is surreal. Really metadata slots are for rollbacks, and should have a consistent view of the partition table at all times. The only exception is during OTAs, where when going from A->B, both slots share "a" partitions and extents, but differ on "b" due to resizing.

> Basically, would this be supported:
> 1) Device booting normally with slot a
> 2) Switch to slot b, flash image, fails to boot
> 3) Switch back to slot a, and it boots normally again
> 
> Assuming the original image and the image flashed in 2) have a different layout.
> 
> Alternatively thought as, like we've been talking about supporting 'a' vs 'b' vs 'all' for slot override, we should only update the metadata for the given slot(s) that fastboot is set to update.

I agree, I broke this when I changed update-super. On the other hand the merging code was pretty flawed, it did not handle changes to groups or block devices. I think the way to fix it is to take the new metadata, and add an "ImportPartitions(<old-metadata>)" function in MetadataBuilder. Then it could figure out whether the old table is compatible with the new metadata format, and error if it's not. At that point fastbootd could wipe if requested, or error if not.

Patch-set: 1
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>