Update patch set 4
Patch Set 4:

Do we write back the adjusted partitions names and remove the flag after the first boot?  I don't see that happening; unless I missed it.

We do need to do that before a subsequent OTA, right?  At least we would need to in the case of growing the super partition to encompass both slots.  I guess we don't if we're going to have separate super partitions for each slot, but that doesn't sound to be the plan.

Patch-set: 4
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>