Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+2
> 
> I just worry that if it's not updated, then we might encounter some unplanned situations in the future.  We should make sure that we have a good plan to test these combinations.

I think the risk is reading the wrong slot: if we've booted after update #2 in the example above, but we read metadata off system_b instead of system_a, things will look weird. update_engine and init should never be doing that, they always use the metadata in the current slot.

fastbootd is a problem though, imagine system_a is upgraded and system_b is not. (This could happen even if we auto-upgraded the metadata on boot, if you skip boot.) In this case, if you stay on "a" and enter fastbootd, everything looks upgraded. If you switch to "b" and enter fastbootd, you're missing half the super partition. The problem is similar if system_b is the initial data and system_a is pre-dynamic-partitions.

When both partitions have metadata, I think fastbootd should just keep them in sync, and auto-upgrade if needed. But I'm not sure about when system_b is dynamic but system_a is still a valid, bootable, non-dynamic slot. Maybe fastbootd should not do anything in this case, or should have a separate command to upgrade.

Patch-set: 4
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0