Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4: Code-Review+2
> > 
> > I just worry that if it's not updated, then we might encounter some unplanned situations in the future.  We should make sure that we have a good plan to test these combinations.
> 
> I think the risk is reading the wrong slot: if we've booted after update #2 in the example above, but we read metadata off system_b instead of system_a, things will look weird. update_engine and init should never be doing that, they always use the metadata in the current slot.

Indeed.  This sounds fine; I'm just emphasizing that we should have some tests to ensure this remains the case.

> 
> fastbootd is a problem though, imagine system_a is upgraded and system_b is not. (This could happen even if we auto-upgraded the metadata on boot, if you skip boot.) In this case, if you stay on "a" and enter fastbootd, everything looks upgraded. If you switch to "b" and enter fastbootd, you're missing half the super partition. The problem is similar if system_b is the initial data and system_a is pre-dynamic-partitions.
> 
> When both partitions have metadata, I think fastbootd should just keep them in sync, and auto-upgrade if needed. But I'm not sure about when system_b is dynamic but system_a is still a valid, bootable, non-dynamic slot. Maybe fastbootd should not do anything in this case, or should have a separate command to upgrade.

`fastboot flashall` should be fine; I think it's sane to always assume that the user means to upgrade to logical partitions during such a flash, if they're already half upgraded.

Flashing an individual physical partition when half upgraded seems like a bit of a corner case.  Maybe worth showing a warning, but maybe okay to allow too; since that won't fundamentally cause any issues and we don't have the ability to upgrade the rest of the physical partitions to logical if we're only flashing a single one.

As a side note, is there anything that we're doing to prevent users from flashing a physical partition on retrofit devices?  Aka prevent flashing the physical system_a, which is actually a super partition?  We'd need to block this in bootloader fastboot especially too.

Patch-set: 4
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>