Update patch set 3
Patch Set 3: Code-Review+1

> Patch Set 3: Verified+1 Presubmit-Ready+1
> 
> I timed this and it's a negligible increase (100us) for the additional chroot and move mounts, which is pretty much expected.  I think it's the right idea moving forward.
> 
> (I also timed freeing the ramdisk from memory and that's 3ms, which is a reasonable trade off too imo)

I think part of the 3ms problem occurs because of the inherent overhead in opendir, it is quite expensive. I switched to syscall(__NR_getdents64, fd, buff[index], sizeof(buff[0]), 0) in llkd and given the wide swath of directories (/proc/*) it was requesting for and more than a 10 fold improvement in that use case. Granted, it's not recursing through directories like you are here, but it also is not allocating a large buffer and context for each and stitching/caching in a complex fashion.

I only optimized in llkd because it is mlocked, as memory allocation in opendir was problematic. The gain in speed was a nice surprise.

_after_ the dust has settled on all of this, it may be worth checking if all the opendir calls in init can be optimized for init's use cases; or if opendir itself could be optimized to work better. I understand saving 2ms (WAG) is not going to impact much. Or, maybe std::filesystem API (?) is more efficient so it may be a simple change?

Patch-set: 3
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1