Update patch set 1
Patch Set 1: Presubmit-Ready+1

At the end of the day, these chained partitions all need to be verified in the bootloader, so this check isn't providing any actual security, right?  This is all just ownership semantics that are in question here?

And with regards to the ownership semantics, it's the case that with 1 vbmeta partition that's vendor owned, vendors already own all of the partitions.  Being able to move to a separate vbmeta (vendor owned) and vbmeta_system (for mainline) moves us away from that.

(Ignoring metadata at the end of partitions).

Those are my thoughts at least.  I'm pretty confident this is fine too.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1