Update patch set 9
Patch Set 9: Presubmit-Ready+1

> Patch Set 9:
> 
> lgtm, but I'm still wondering why not 0444?

Sorry somehow ignored your previous comment.

Initially I thought it may be useful to rewrite the file when the system property is updated. So that they are always in sync, and we can easily test / benchmark different routines.

But should be easy to change when we really do that. Use 444 for now.

Patch-set: 9
Reviewer: Gerrit User 1307957 <1307957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1