proof of concept: create all block devices in first stage mount
First stage mount started off rather simply: we were going to find a
small set of partitions and create their /dev devices along with their
dm-X devices for verity.  This has subsequently been complicated as
with AVB, there are a list of partitions that may contain AVB metadata
that need their /dev devices to be created, and further complicated
with dm-linear which may require any number of underlying physical
partitons to have their /dev devices created.

All in all, creating a handful of /dev devices is rather cheap, and it
seems worth it at this point to create all of them rather than the
complex logic and multiple inputs required to create just a
not-small-anymore subset of them.

Bug: 117933812
Test: not yet
Change-Id: I165eb88eb3ab140d02406de12fae0c9cd6777c8c