Update patch set 2
Patch Set 2:

I don't have a great answer here.  We have an internal bug tracking moving the check about whether or not something should be logged earlier, but the problem there is that if we move this check too early, that the timestamp of the log messages will be skewed.  And if we can't move that check to the header, then we can't provide a default other than VERBOSE when checking logging later on (since the later check would be generic in liblog and wouldn't be able to know if NDEBUG was set in the calling file).

I don't have a better solution than CL744768 to offer in the short term.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>