Create BoringSSL test program
NIAP certification requires that all cryptographic functions
undergo a self-test during startup to demonstrate correct
operation. This new program, boringssl_self_test, performs
this check.

If the check fails, the new program needs to prevent the device
from booting. To this end, it will cause the device to boot into
the bootloader.

Test: Built for walleye. When run it reports that BoringSSL's
      self test is successful. Forcing the self check to fail
      causes the device to reboot into the boot loader.

Change-Id: I07a5dc73a314502c87de566bb26f4d73499d2675