Create BoringSSL test program, run during boot
NIAP certification requires that all cryptographic functions
undergo a self-test during startup to demonstrate correct
operation. This new program, boringssl_self_test, performs
this check.

If the check fails, the new program needs to prevent the device
from booting. To this end, it will cause the device to boot into
the bootloader.

Bug: 119826244
Test: Built for walleye. After device booted examined dmesg and
      observed logs from init showing that the new service did
      start and exit successfully. Further, the service finished
      before the device booted past the boot animation. Finally,
      when BoringSSL is built to fail its self check the device
      will stop during a normal boot and enter the bootloader.

Change-Id: I07a5dc73a314502c87de566bb26f4d73499d2675