Add ANDROID_RUNTIME environment variable
This change adds an ANDROID_RUNTIME environment variable
to indicate the root of the Android runtime. This can be
used in place of ANDROID_SYSTEM to locate files when they
move inside the runtime APEX.

The alternative would be to hardcode /apex/com.android.runtime
(which breaks situations like host linux which cannot provide
the expected file structure), to use preprocessor or global
variables to modify the path used at runtime.

Bug: 119935277
Test: build / boot
Change-Id: Ic3b1ae3e3c98eea7d7c59e514ce62994679ab7b7