fs_mgr: fastboot: init: overlay: use alternate super for backing storage
When we have multiple physical super partitions, it is prudent to
take the alternate super partition in its entirety than to contruct
a logical partition out of the current super partition.

init add support for picking up partitions if specified by
fs_mgr_overlayfs_required_devices().

fastboot call fs_mgr_overlayfs_teardown for update-super command.

For specific teardown operations, if we mount scratch, umount it
otherwise it interferes with fastboot.

Disable overrides if we are booting up in recovery or fastbootd.
Check for androidboot.force_normal_boot=1 to permit overrides.

SideEffect: alternate boot partitions destroyed for the sake of debug
Test: adb-remount-test.sh
Bug: 119885423
Bug: 120034852
Change-Id: I368d3a5619f69de3b0d1fcad73c4b42d25d0d969