get the ip address info for all DNS services
bug: 119490749

This CL adds functionality to the class ResolvedService where the ip
address associated with a resolved DNS service is recorded.

It also avoids connecting to the device unless it is the Things-related
DNS service.

Next step is to add some kind of interface in other parts of adb code
to retrieve these IP addresses.

Change-Id: I46895a7a5bf5660f524c7323a9454f1e2c7d385c