Update patch set 5
Patch Set 5:

> Would we want this to be a part of init directly? ... this is simple enough that it might be worth just including in.

I've been back and forth on this.

Having this check be a part of init seems a little "magical", and (at least to me) it does not seem like something init should be responsible for. 

There is also the question of startup performance. Currently the check can execute concurrently with startup. Experiments on a Pixel 2 show that the check takes on average 180 ms to complete, however there were instances that took as much as 1.3 seconds. (This must be due to the load at at the time it was running at startup). However, overall the checks did not appear to have an impact on boot times. Would running the check as a part of 2nd state init force it to gate startup until complete? If so, the checks should negatively impact startup times. It may not be as much as ~1 sec, but it may be non-trivial.

Patch-set: 5
Reviewer: Gerrit User 1398097 <1398097@85c56323-6fa9-3386-8a01-6480fb634889>