fastboot: optional secondary partitions on retrofit devices
If secondary partitions is missing, or not matching the primary
logical type, skip them.  Warn user when skipped since
--skip-secondary was not specified.  This logic will trigger only
under retrofit or dual super situation.

This is mainly a flashing error mitigation strategy, as for example
system_b might not flash because it is physical, when system_a is
logical, but boot_b will flash if always on a physical partition
despite the mismatch for system.  Without this change flashall
will error out with system_b and not complete the flashing of the
a slot.

Test: fastboot flashall (where a and b match, or do not match)
Bug: 120034852
Change-Id: I6830343e6c298b90173e94a0e1070e08c2cf1c42