Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Would we want this to be a part of init directly? ... this is simple enough that it might be worth just including in.
> 
> I've been back and forth on this.
> 
> Having this check be a part of init seems a little "magical", and (at least to me) it does not seem like something init should be responsible for. 

Ya, it's a rather gray area.  In one way, it'd be nice if init were generic and all of the Android specific behavior were factored out into separate executables like this, but I think we're long past that point as init is directly responsible for a lot of Android specific SELinux and file system set up.  Adding this wouldn't seem to be that unreasonable.

> 
> There is also the question of startup performance. Currently the check can execute concurrently with startup. Experiments on a Pixel 2 show that the check takes on average 180 ms to complete, however there were instances that took as much as 1.3 seconds. (This must be due to the load at at the time it was running at startup). However, overall the checks did not appear to have an impact on boot times. Would running the check as a part of 2nd state init force it to gate startup until complete? If so, the checks should negatively impact startup times. It may not be as much as ~1 sec, but it may be non-trivial.

We can always fork() this code from init directly and let it run in parallel with the rest of init.

Patch-set: 5
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>