Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> If init exits, the kernel reboots.

However, if the forked child that runs the BoringSSL check hits a failure and cannot reboot, exiting will not cause init to exit.

If it is possible for init to not have SYS_BOOT permission, should the BoringSSL test be forked? Or, maybe, should the fork be dependent on having SYS_BOOT?

Patch-set: 5
Reviewer: Gerrit User 1398097 <1398097@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0