fastboot: do not die if ANDROID_PRODUCT_OUT undefined
When checking for existence of "super_empty.img" to determine if
flash image product set is meant for logical partitions, we die if
ANDROID_PRODUCT_OUT environment is unset or empty.  This check
is done before we look at the flash image name to determine if it
is a candidate to look at the logical metadata.

Do not die, return empty string to find_item_given_name() and let
caller decide.

Add O_CLOEXEC to open() as per security standards.

Test: compile
Bug: 120041144
Change-Id: I43f124015f9d26c79a0feb9123522432fe937343