Update patch set 2
Patch Set 2:

> Patch Set 1: Code-Review-2
> 
> (wait a minute, sorry).
> 
> I think this breaks the error we print if someone is trying to flash a logical partition using bootloader. Fastboot will happily go ahead with this.

It's okay if it doesn't work in this scenario (flashing a specific image with no OUT dir) since the bootloader has its own protection as a fallback.

> I think the key decision to decide if the partition needs userspace fastboot should be based on something other than the presence of super_empty.img / ANDROID_PRODUCT_OUT ...

We don't know what partitions are dynamic in the bootloader. (We don't even know in fastbootd, since the build could be introducing a new one.) The only thing that has that information is super_empty.img.

In a situation like this where we're flashing a one-off image, and we don't have any other information, I think we should just return false in should_flash_in_userspace(). Sticking the ANDROID_PRODUCT_OUT check there would fix it.

Patch-set: 2
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>