Update patch set 3
Patch Set 3: Code-Review+2

> Patch Set 1: -Autosubmit
> 
> Need to add a change to the bootloader and to hard code known logical partitions when device tells us it is logical.

ok, (The retrofit bootloader does that already to protect against overwriting system, so it's moot there.).

Confirmed with @hridya too that the bootloader has a built-in check and will not flash the physical partition *ever*. This kinda breaks the "--force" use cases for retrofits, but that's ok and can be fixed right after.

So, unblocking lab for now so TH can proceed is better.

Patch-set: 3
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2