adbd: only register dnsservices via explicit API
bug: 119490749

For privacy/security reasons, let's not broadcast the adb secure
services on startup automatically, and instead leave that up to
the rest of the adbd code somehow.

Instead, this CL adds an API in daemon/mdns.h that lets the user
control when registration happens, potentially only doing so if
the developer option is enabled or something.

Change-Id: Idc994a59ef9e5d48f08796c21989883497e19ef8