Update patch set 4
Patch Set 4:

What happens to OTAs in this case?  Are people allowed to stay on the OTA train with dm-verity disabled?  If so, this would presumably break that, since update_engine wouldn't be able to use the alternate slot for the OTA?

Patch-set: 4
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>