Update patch set 7
Patch Set 7:

Silly question: do we teardown scratch of 'system' when we flash a new 'system' ?  We should, right?  I just don't remember seeing the code for this go by.

I'm still a bit nervous about anything that involves margins.  This essentially means that if I flash a vendor that is the same size as the original one, that we don't tear down the old scratch with system, but if I flash a vendor that's slightly larger than the original one, that we do tear down the old scratch with system?

Patch-set: 7
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>