Update patch set 7
Patch Set 7:

If we already tear down scratch in almost all cases, why not tear it down in every case?  We wouldn't need any of this margin in that case, right?

In other words, what cases does this margin help us with?  Flashing a same sized system.img without flashing a new super?  (And do we have the code to tear down just the system portion of the super in this case too? We'd want to do that even if we keep super around for its overlay of vendor or other partitions)

Patch-set: 7
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>